refactor: migrate TechSkillsCard to TypeScript with skill editing and backend sync
